+++ packages/fx-core/package.json
@@ -64,7 +64,7 @@
     "test:projcheck": "nyc mocha \"tests/common/projectTypeChecker.test.ts\"",
     "test:telemetry": "nyc mocha \"tests/common/telemetry.test.ts\"",
     "test:stringUtils": "nyc mocha \"tests/common/stringUtils.test.ts\"",
-    "test:genUtils": "nyc mocha \"tests/component/generator/genUtils.test.ts\"",
+    "test:generatorUtils": "nyc mocha \"tests/component/generatorUtils.test.ts\"",
     "clean": "rm -rf build",
     "prebuild": "npm run gen:cli",
     "build": "rimraf build && npx tsc -p ./",
