+++ packages/fx-core/src/component/generator/utils.ts
@@ -399,53 +399,3 @@ export function isApiLimitError(error: Error): boolean {
     error.response?.headers?.["x-ratelimit-remaining"] === "0"
   );
 }
-
-export async function fetchAndUnzip(
-  component: string,
-  zipUrl: string,
-  targetDir: string,
-  skipRootFolder = true
-): Promise<Result<undefined, FxError>> {
-  let zip: AdmZip;
-  let response: any;
-  try {
-    zip = await fetchZipFromUrl(zipUrl);
-  } catch (e: any) {
-    return err(new AccessGithubError(zipUrl, component, e));
-  }
-  if (!zip) {
-    return err(
-      new AccessGithubError(
-        zipUrl,
-        component,
-        new Error(
-          `Failed to fetch GitHub URL: ${response.status as string} ${
-            response.statusText as string
-          }`
-        )
-      )
-    );
-  }
-  const entries = zip.getEntries();
-  let rootFolderName = "";
-  for (const entry of entries) {
-    const entryName: string = entry.entryName;
-    if (skipRootFolder && !rootFolderName) {
-      rootFolderName = entryName;
-      continue;
-    }
-    const rawEntryData: Buffer = entry.getData();
-    const entryData: string | Buffer = rawEntryData;
-    const targetPath = path.join(targetDir, entryName.replace(rootFolderName, ""));
-    try {
-      if (entry.isDirectory) {
-        await fs.ensureDir(targetPath);
-      } else {
-        await fs.writeFile(targetPath, entryData);
-      }
-    } catch (error: any) {
-      return err(new ReadFileError(error, component));
-    }
-  }
-  return ok(undefined);
-}
